+++ tests/unit/server/cli/search_engine/test_reindex.py
@@ -24,11 +24,11 @@ class TestCliServerSearchEngineReindex:
     # Instead of do that we can (once we move away from asynchronous requests) use regular factories here and improve
     # the test.
     def test_reindex(self, cli_runner: CliRunner, cli: Typer):
-        result = cli_runner.invoke(cli, "server search-engine reindex")
+        result = cli_runner.invoke(cli, "search-engine reindex")
 
-        assert result.exit_code == 0
+        assert result.exit_code == 0, result.output
 
     def test_reindex_with_nonexistent_dataset_id(self, cli_runner: CliRunner, cli: Typer):
-        result = cli_runner.invoke(cli, f"server search-engine reindex --feedback-dataset-id {uuid4()}")
+        result = cli_runner.invoke(cli, f"search-engine reindex --feedback-dataset-id {uuid4()}")
 
         assert result.exit_code == 1